+++ src/cli_test.rs
@@ -5,105 +5,135 @@ use types::{CliArgs, GlobalConfig};
 
 #[test]
 fn run_empty_task() {
-    run(CliArgs {
-        build_file: "bad.toml".to_string(),
-        task: "empty".to_string(),
-        log_level: "error".to_string(),
-        cwd: None,
-        env: None,
-        disable_workspace: false,
-        disable_check_for_updates: true,
-        print_only: false,
-        list_all_steps: false,
-        experimental: false,
-    });
+    let global_config = GlobalConfig::new();
+
+    run(
+        CliArgs {
+            build_file: "bad.toml".to_string(),
+            task: "empty".to_string(),
+            log_level: "error".to_string(),
+            cwd: None,
+            env: None,
+            disable_workspace: false,
+            disable_check_for_updates: true,
+            print_only: false,
+            list_all_steps: false,
+            experimental: false,
+        },
+        &global_config,
+    );
 }
 
 #[test]
 fn print_empty_task() {
-    run(CliArgs {
-        build_file: "bad.toml".to_string(),
-        task: "empty".to_string(),
-        log_level: "error".to_string(),
-        cwd: None,
-        env: None,
-        disable_workspace: false,
-        disable_check_for_updates: true,
-        print_only: true,
-        list_all_steps: false,
-        experimental: false,
-    });
+    let global_config = GlobalConfig::new();
+
+    run(
+        CliArgs {
+            build_file: "bad.toml".to_string(),
+            task: "empty".to_string(),
+            log_level: "error".to_string(),
+            cwd: None,
+            env: None,
+            disable_workspace: false,
+            disable_check_for_updates: true,
+            print_only: true,
+            list_all_steps: false,
+            experimental: false,
+        },
+        &global_config,
+    );
 }
 
 #[test]
 fn list_empty_task() {
-    run(CliArgs {
-        build_file: "bad.toml".to_string(),
-        task: "empty".to_string(),
-        log_level: "error".to_string(),
-        cwd: None,
-        env: None,
-        disable_workspace: false,
-        disable_check_for_updates: true,
-        print_only: false,
-        list_all_steps: true,
-        experimental: false,
-    });
+    let global_config = GlobalConfig::new();
+
+    run(
+        CliArgs {
+            build_file: "bad.toml".to_string(),
+            task: "empty".to_string(),
+            log_level: "error".to_string(),
+            cwd: None,
+            env: None,
+            disable_workspace: false,
+            disable_check_for_updates: true,
+            print_only: false,
+            list_all_steps: true,
+            experimental: false,
+        },
+        &global_config,
+    );
 }
 
 #[test]
 fn run_file_and_task() {
-    run(CliArgs {
-        build_file: "./examples/dependencies.toml".to_string(),
-        task: "A".to_string(),
-        log_level: "error".to_string(),
-        cwd: None,
-        env: None,
-        disable_workspace: false,
-        disable_check_for_updates: true,
-        print_only: false,
-        list_all_steps: false,
-        experimental: false,
-    });
+    let global_config = GlobalConfig::new();
+
+    run(
+        CliArgs {
+            build_file: "./examples/dependencies.toml".to_string(),
+            task: "A".to_string(),
+            log_level: "error".to_string(),
+            cwd: None,
+            env: None,
+            disable_workspace: false,
+            disable_check_for_updates: true,
+            print_only: false,
+            list_all_steps: false,
+            experimental: false,
+        },
+        &global_config,
+    );
 }
 
 #[test]
 fn run_cwd_with_file() {
+    let global_config = GlobalConfig::new();
+
     let directory = Path::new("./examples");
     assert!(env::set_current_dir(&directory).is_ok());
 
-    run(CliArgs {
-        build_file: "./examples/dependencies.toml".to_string(),
-        task: "A".to_string(),
-        log_level: "error".to_string(),
-        cwd: Some("..".to_string()),
-        env: None,
-        disable_workspace: false,
-        disable_check_for_updates: true,
-        print_only: false,
-        list_all_steps: false,
-        experimental: false,
-    });
+    run(
+        CliArgs {
+            build_file: "./examples/dependencies.toml".to_string(),
+            task: "A".to_string(),
+            log_level: "error".to_string(),
+            cwd: Some("..".to_string()),
+            env: None,
+            disable_workspace: false,
+            disable_check_for_updates: true,
+            print_only: false,
+            list_all_steps: false,
+            experimental: false,
+        },
+        &global_config,
+    );
 }
 
 #[test]
 #[should_panic]
 fn run_cwd_task_not_found() {
+    let global_config = GlobalConfig::new();
+
     let directory = Path::new("./examples");
     assert!(env::set_current_dir(&directory).is_ok());
 
-    run(CliArgs {
-        build_file: "./dependencies.toml".to_string(),
-        task: "A".to_string(),
-        log_level: "error".to_string(),
-        cwd: Some("..".to_string()),
-        env: None,
-        disable_workspace: false,
-        disable_check_for_updates: true,
-        print_only: false,
-        list_all_steps: false,
-        experimental: false,
-    });
+    run(
+        CliArgs {
+            build_file: "./dependencies.toml".to_string(),
+            task: "A".to_string(),
+            log_level: "error".to_string(),
+            cwd: Some("..".to_string()),
+            env: None,
+            disable_workspace: false,
+            disable_check_for_updates: true,
+            print_only: false,
+            list_all_steps: false,
+            experimental: false,
+        },
+        &global_config,
+    );
 }
 
 #[test]
@@ -120,7 +150,7 @@ fn run_for_args_bad_subcommand() {
 
     let matches = app.get_matches_from(vec!["bad"]);
 
-    run_for_args(matches);
+    run_for_args(matches, &global_config);
 }
 
 #[test]
@@ -139,7 +169,7 @@ fn run_for_args_valid() {
         "error",
     ]);
 
-    run_for_args(matches);
+    run_for_args(matches, &global_config);
 }
 
 #[test]
@@ -151,7 +181,7 @@ fn run_for_args_with_global_config() {
 
     let matches = app.get_matches_from(vec!["cargo", "make"]);
 
-    run_for_args(matches);
+    run_for_args(matches, &global_config);
 }
 
 #[test]
@@ -171,7 +201,7 @@ fn run_for_args_log_level_override() {
         "-v",
     ]);
 
-    run_for_args(matches);
+    run_for_args(matches, &global_config);
 }
 
 #[test]
@@ -198,7 +228,7 @@ fn run_for_args_set_env() {
         "empty",
     ]);
 
-    run_for_args(matches);
+    run_for_args(matches, &global_config);
 
     assert_eq!(env::var("ENV1_TEST").unwrap(), "TEST1");
     assert_eq!(env::var("ENV2_TEST").unwrap(), "TEST2");
@@ -224,5 +254,5 @@ fn run_for_args_print_only() {
         "--experimental",
     ]);
 
-    run_for_args(matches);
+    run_for_args(matches, &global_config);
 }