+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.10.0 (TODO)
 
 * Global configuration file to define default log level, task name (#64)
+* Check for update minimum interval definition in global configuration (#68)
 * Reformat changelog (#66)
 
 ### v0.9.5 (2018-01-22)